Update README.md
In this paper, a variant of the particle swarm optimisation (PSO) algorithm is introduced with heterogeneous behaviour and a new dynamic multi-swarm topological structure. The new topological structure enables the algorithm to have more control over the interaction and information exchange between the particles to reduce the loss of diversity and avoid premature convergence. In the new algorithm, the population is initially divided into two sub-populations, first sub-population is further divided into sub-swarms that are formed using the introduced topological structure. The particles of sub-swarms are guided using heterogeneous behaviour by selecting various exemplars. The second sub-population employs the classical PSO search with local and global information to simulate a homogenous behaviour. There is information flow between the two subpopulations. The algorithm was tested on the CEC2005 and CEC2017 test suites with comparison against various state-of the-art PSO variants and other state-of-the-art meta-heuristics. The experimental results show that for the two test suites, the proposed algorithm outperformed the majority of the state-of the-art algorithms on most problems.